+++ Cargo.toml
@@ -22,8 +22,8 @@ default = [
     "bevy_gltf",
     "bevy_wgpu",
     "bevy_winit",
+    "bevy_dynload",
     "render",
-    "dynamic_plugins",
     "png",
     "hdr",
     "mp3",
@@ -31,11 +31,7 @@ default = [
 ]
 profiler = ["bevy_ecs/profiler", "bevy_diagnostic/profiler"]
 wgpu_trace = ["bevy_wgpu/trace"]
-dynamic_plugins = [
-    "bevy_core/dynamic_plugins",
-    "bevy_app/dynamic_plugins",
-    "bevy_type_registry/dynamic_plugins",
-]
+
 # Rendering support
 render = ["bevy_pbr", "bevy_render", "bevy_sprite", "bevy_text", "bevy_ui"]
 # Image format support for texture loading (PNG and HDR are enabled by default)
@@ -79,6 +75,7 @@ bevy_audio = { path = "crates/bevy_audio", optional = true, version = "0.2.1" }
 bevy_gltf = { path = "crates/bevy_gltf", optional = true, version = "0.2.1" }
 bevy_pbr = { path = "crates/bevy_pbr", optional = true, version = "0.2.1" }
 bevy_render = { path = "crates/bevy_render", optional = true, version = "0.2.1" }
+bevy_dynload = { path = "crates/bevy_dynload", optional = true, version = "0.2.1" }
 bevy_sprite = { path = "crates/bevy_sprite", optional = true, version = "0.2.1" }
 bevy_text = { path = "crates/bevy_text", optional = true, version = "0.2.1" }
 bevy_ui = { path = "crates/bevy_ui", optional = true, version = "0.2.1" }
