+++ crates/bevy_app/src/plugin.rs
@@ -1,6 +1,4 @@
 use crate::AppBuilder;
-#[cfg(feature = "dynamic_plugins")]
-use libloading::{Library, Symbol};
 use std::any::Any;
 
 /// A collection of Bevy App logic and configuration
@@ -14,15 +12,3 @@ pub trait Plugin: Any + Send + Sync {
 }
 
 pub type CreatePlugin = unsafe fn() -> *mut dyn Plugin;
-
-#[cfg(feature = "dynamic_plugins")]
-/// Dynamically links a plugin a the given path. The plugin must export the [CreatePlugin] function.
-pub fn dynamically_load_plugin(path: &str) -> (Library, Box<dyn Plugin>) {
-    let lib = Library::new(path).unwrap();
-
-    unsafe {
-        let func: Symbol<CreatePlugin> = lib.get(b"_create_plugin").unwrap();
-        let plugin = Box::from_raw(func());
-        (lib, plugin)
-    }
-}
