+++ crates/bevy_core/Cargo.toml
@@ -12,11 +12,6 @@ repository = "https://github.com/bevyengine/bevy"
 license = "MIT"
 keywords = ["bevy"]
 
-[features]
-dynamic_plugins = [
-    "bevy_app/dynamic_plugins",
-    "bevy_type_registry/dynamic_plugins",
-]
 
 [dependencies]
 bevy_app = { path = "../bevy_app", version = "0.2.1" }
