+++ crates/bevy_dynload/src/lib.rs
@@ -1,0 +1,3 @@
+mod loader;
+
+pub use loader::*;
