+++ src/prelude.rs
@@ -21,3 +21,6 @@ pub use crate::text::prelude::*;
 
 #[cfg(feature = "bevy_ui")]
 pub use crate::ui::prelude::*;
+
+#[cfg(feature = "bevy_dynload")]
+pub use crate::dynload::*;
